requirements.txt
requests==2.31.0
auth0-python==4.4.0
invoke==2.2.0